% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/symAlgebra.R
\docType{class}
\name{sAlg-class}
\alias{sAlg}
\alias{sAlg-class}
\title{Class that represents a symbolic expression}
\description{
Class that represents a symbolic expression
}
\section{Slots}{

\describe{
\item{\code{pointer}}{Pointer to a Symbolic object from SymbolicC++}
}}

